+++ CHANGELOG.md
@@ -5,8 +5,10 @@
   - Section 2: JavaScript Language Fundamentals
     - General Loops
     - A Look At The Window Object
+    - Block Scope With let & const
 - An example with loops for General Loops was added.
 - An example with the window object for A Look At The Window Object was added.
+- An example with scopes for Block Scope With let & const was added.
 
 ## August 5, 2020
 - Modern JavaScript From The Beginning (Udemy)